Flytt oppgavebenkstats inn i egen fil
