+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/utils/oppgavebenkStats.ts
@@ -1,0 +1,9 @@
+export interface OppgavebenkStats {
+  antallOppgavelistaOppgaver: number
+  antallMinOppgavelisteOppgaver: number
+}
+
+export const initalOppgavebenkStats: OppgavebenkStats = {
+  antallOppgavelistaOppgaver: 0,
+  antallMinOppgavelisteOppgaver: 0,
+}
